requirements.txt
pydantic
langgraph
requests
Pillow
piexif
pillow-heif
numpy
